+++ public.ssl/admin/users/admin_user_download_cert.php
@@ -12,7 +12,7 @@ if (!$me->authenticate()) {
     exit();
 }
 
-if (!$me->isAdmin()) {
+if (!$me->isAnyAdmin()) {
     $_SESSION["error"] = "Accès refusé";
     header("Location: " . WEBSITE_SSL);
     exit();