+++ public.ssl/modules/helios/index.php
@@ -64,7 +64,7 @@ if ($droit->isSuperAdmin($initData->userInfo)) {
     } else {
         $heliosTransactionsListe->setAuthority($initData->userInfo['authority_id']);
     }
-} elseif ($droit->isAdmin($initData->userInfo)) {
+} elseif ($droit->isAnyAdmin($initData->userInfo)) {
     $heliosTransactionsListe->setAuthority($initData->userInfo['authority_id']);
 } else {
     $serviceUser = new ServiceUser(DatabasePool::getInstance());